+++ ChangeLog.md
@@ -5,5 +5,6 @@
 
 ## 0.1.1.0
 - follow `INLINE` pragmas when adding loopbreakers
+- support local declarations (`let` and `where` bindings)
 
 ## Unreleased changes